requirements.txt
streamlit
PyPDF2
scikit-learn
pandas
nltk